+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## <WIP>
 
 - Add support for changes outside of Terraform
+- Add support for output changes of Terraform
 
 ## 0.3.3